#3 feat: create function to get pages from sheet
